+++ packages/auth-construct/src/types.ts
@@ -1,5 +1,7 @@
 import { aws_cognito as cognito } from 'aws-cdk-lib';
 import { AuthUserAttribute } from './attributes.js';
+import { triggerEvents } from './trigger_events.js';
+
 /**
  * Email login options.
  *
@@ -68,6 +70,11 @@ export type ExternalProviders = {
 };
 
 /**
+ * Union type of all supported auth trigger events
+ */
+export type TriggerEvent = (typeof triggerEvents)[number];
+
+/**
  * Input props for the AmplifyAuth construct
  */
 export type AuthProps = {
@@ -80,4 +87,12 @@ export type AuthProps = {
    * Multifactor Authentication settings
    */
   multifactor?: MFA;
+  /**
+   * Determined how a user is able to recover their account by setting the account recovery setting.
+   *
+   * If no setting is provided, a default will be set based on the enabled login methods.
+   * When email and phone login methods are both enabled, email will be the default recovery method.
+   * If only email or phone are enabled, they will be the default recovery methods.
+   */
+  accountRecovery?: cognito.AccountRecovery;
 };