+++ packages/plugin-types/src/auth_resources.ts
@@ -5,6 +5,7 @@ import {
   IUserPool,
   IUserPoolClient,
 } from 'aws-cdk-lib/aws-cognito';
+
 /**
  * Accessible Cfn resources from the Auth construct, which were generated during
  * initialization.