+++ scripts/get_unit_test_dir_list.ts
@@ -1,0 +1,9 @@
+import { glob } from 'glob';
+import * as path from 'path';
+
+let result = await glob('packages/*');
+result = result.filter((result) => !result.includes('integration-tests'));
+result.push(
+  path.join('packages', 'integration-tests', 'lib', 'test-in-memory')
+);
+console.log(result.join(' '));