+++ packages/function-construct/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @aws-amplify/function-construct-alpha
 
+## 0.1.1-alpha.2
+
+### Patch Changes
+
+- 3c36ace: Implement UserPool trigger config
+
 ## 0.1.1-alpha.1
 
 ### Patch Changes